+++ include/loot/game_interface.h
@@ -103,27 +103,42 @@ public:
 
   /**
    * @brief Parses plugins and loads their data.
-   * @details Any previously-loaded plugin data is discarded when this function
-   *          is called.
+   * @details If a given plugin filename (or one that is case-insensitively
+   *          equal) has already been loaded, its previously-loaded data
+   *          data is discarded, invalidating any existing shared pointers to
+   *          that plugin's PluginInterface object.
+   *
+   *          If the game is Morrowind, OpenMW or Starfield, it's only valid to
+   *          fully load a plugin if its masters are already loaded or included
+   *          in the same input vector.
    * @param pluginPaths
    *        The plugin paths to load. Relative paths are resolved relative to
    *        the game's plugins directory, while absolute paths are used as
    *        given. Each plugin filename must be unique within the vector.
    * @param loadHeadersOnly
    *        If true, only the plugins' headers are loaded. If false, all records
-   *        in the plugins are parsed, apart from the main master file if it has
-   *        been identified by a previous call to ``IdentifyMainMasterFile()``.
+   *        in the plugins are parsed.
    */
   virtual void LoadPlugins(
       const std::vector<std::filesystem::path>& pluginPaths,
       bool loadHeadersOnly) = 0;
 
   /**
+   * @brief Clears the plugins loaded by previous calls to `LoadPlugins()`.
+   * @details This invalidates any PluginInterface pointers retrieved using
+   *          `GetPlugin()` or `GetLoadedPlugins()`.
+   */
+  virtual void ClearLoadedPlugins() = 0;
+
+  /**
    * @brief Get data for a loaded plugin.
    * @param  pluginName
    *         The filename of the plugin to get data for.
    * @returns A shared pointer to a const PluginInterface implementation. The
-   *          pointer is null if the given plugin has not been loaded.
+   *          pointer is null if the given plugin has not been loaded. The
+   *          pointer remains valid until the `ClearLoadedPlugins()` function
+   *          is called, this GameInterface is destroyed, or until a plugin with
+   *          a case-insensitively equal filename is loaded.
    */
   virtual const PluginInterface* GetPlugin(
       const std::string& pluginName) const = 0;
@@ -131,9 +146,10 @@ public:
   /**
    * @brief Get a set of const references to all loaded plugins' PluginInterface
    *        objects.
-   * @returns A set of const PluginInterface references. The references remain
-   *          valid until the ``LoadPlugins()`` or ``SortPlugins()`` functions
-   *          are next called or this GameInterface is destroyed.
+   * @returns A set of shared pointers to const PluginInterface. The pointers
+   *          remain valid until the `ClearLoadedPlugins()` function is called,
+   *          this GameInterface is destroyed, or until a plugin with a
+   *          case-insensitively equal filename is loaded.
    */
   virtual std::vector<const PluginInterface*> GetLoadedPlugins() const = 0;
 
@@ -144,16 +160,6 @@ public:
    */
 
   /**
-   *  @brief Identify the game's main master file.
-   *  @details When sorting, LOOT always only loads the headers of the game's
-   *           main master file as a performance optimisation.
-   *
-   *           A relative path is resolved relative to the game's plugins
-   *           directory, while an absolute path is used as given.
-   */
-  virtual void IdentifyMainMasterFile(const std::filesystem::path& masterFile) = 0;
-
-  /**
    *  @brief Calculates a new load order for the game's installed plugins
    *         (including inactive plugins) and outputs the sorted order.
    *  @details Pulls metadata from the masterlist and userlist if they are
@@ -161,15 +167,13 @@ public:
    *           applied to the load order used by the game. This function does
    *           not load or evaluate the masterlist or userlist.
    *  @param pluginPaths
-   *         The plugin paths to sort, in their current load order. Relative
-   *         paths are resolved relative to the game's plugins directory, while
-   *         absolute paths are used as given. Each plugin filename must be
-   *         unique within the vector.
+   *         The plugins to sort, in their current load order. All given plugins
+   *         must have been loaded using `LoadPlugins()`.
    *  @returns A vector of the given plugin filenames in their sorted load
    *           order.
    */
   virtual std::vector<std::string> SortPlugins(
-      const std::vector<std::filesystem::path>& pluginPaths) = 0;
+      const std::vector<std::string>& pluginFilenames) = 0;
 
   /**
    *  @}