+++ src/api/game/game.h
@@ -70,15 +70,15 @@ public:
   void LoadPlugins(const std::vector<std::filesystem::path>& pluginPaths,
                    bool loadHeadersOnly) override;
 
+  void ClearLoadedPlugins() override;
+
   const PluginInterface* GetPlugin(
       const std::string& pluginName) const override;
 
   std::vector<const PluginInterface*> GetLoadedPlugins() const override;
 
-  void IdentifyMainMasterFile(const std::filesystem::path& masterFile) override;
-
   std::vector<std::string> SortPlugins(
-      const std::vector<std::filesystem::path>& pluginPaths) override;
+      const std::vector<std::string>& pluginFilenames) override;
 
   void LoadCurrentLoadOrderState() override;
 
@@ -102,8 +102,6 @@ private:
   LoadOrderHandler loadOrderHandler_;
   std::shared_ptr<ConditionEvaluator> conditionEvaluator_;
   ApiDatabase database_;
-
-  std::filesystem::path masterFilePath_;
 
   std::vector<std::filesystem::path> additionalDataPaths_;
 };