+++ src/tests/api/internals/sorting/plugin_sorting_data_test.h
@@ -46,7 +46,6 @@ protected:
       }
     }
 
-    game.IdentifyMainMasterFile(std::filesystem::u8path(masterFile));
     game.LoadCurrentLoadOrderState();
     game.LoadPlugins(plugins, headersOnly);
   }